+++ runner/result.go
@@ -5,5 +5,5 @@ import "time"
 type Result struct {
 	Log        string
 	TimeTaken  time.Duration
-	MemoryUsed int
+	MemoryUsed uint64
 }
